DataBlockHashIndex: rename data_block_index_format.{h,cc} to data_block_footer.{h,cc}
