+++ table/data_block_footer.cc
@@ -7,7 +7,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "data_block_index_format.h"
+#include "data_block_footer.h"
 
 #include "rocksdb/table.h"
